+++ actix-http/src/body/none.rs
@@ -10,26 +10,30 @@ use super::{BodySize, MessageBody};
 
 /// Body type for responses that forbid payloads.
 ///
-/// Distinct from an empty response which would contain Content-Length header.
+/// Distinct from an empty response which would contain a Content-Length header.
 ///
 /// For an "empty" body, use `()` or `Bytes::new()`.
 #[derive(Debug, Clone, Copy, Default)]
-pub struct None(());
+#[non_exhaustive]
+pub struct None;
 
 impl None {
     /// Constructs new "none" body.
+    #[inline]
     pub fn new() -> Self {
-        None(())
+        None
     }
 }
 
 impl MessageBody for None {
     type Error = Infallible;
 
+    #[inline]
     fn size(&self) -> BodySize {
         BodySize::None
     }
 
+    #[inline]
     fn poll_next(
         self: Pin<&mut Self>,
         _cx: &mut Context<'_>,